device: Fix the wrong order of UUIDs PropertyChanged
The code goes to great length to make sure that the Service Discovery
is finished before sending the NewConnection() message to external
profiles.  In some situations this effort is wasted because the UUIDs
signal is emitted after NewConnection() is sent.

This patch fixes it by making sure that all the properties for that
object are updated before sending the NewConnection() call.